# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MarketPulse API",
    "description": "Yön olasılığı ve piyasa durumu. İşlem yapmaz.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/market/{symbol}": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Market",
        "operationId": "get_market_api_v1_market__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketStateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/{symbol}/candles": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Candles",
        "operationId": "get_candles_api_v1_market__symbol__candles_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Interval",
              "default": "15m"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "List Predictions",
        "operationId": "list_predictions_api_v1_predictions_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Horizon"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "live",
                    "baseline",
                    "backtest"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "resolved",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/{prediction_id}": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Prediction",
        "operationId": "get_prediction_api_v1_predictions__prediction_id__get",
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandleCoverageOut": {
        "properties": {
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "last_close_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Close Time"
          }
        },
        "type": "object",
        "required": [
          "interval",
          "count",
          "last_close_time"
        ],
        "title": "CandleCoverageOut"
      },
      "CandleOut": {
        "properties": {
          "time": {
            "type": "integer",
            "title": "Time"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "type": "number",
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "time",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "CandleOut",
        "description": "lightweight-charts biçimi: `time` saniye cinsinden UNIX zamanı (mum açılışı)."
      },
      "CandlesOut": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "candles": {
            "items": {
              "$ref": "#/components/schemas/CandleOut"
            },
            "type": "array",
            "title": "Candles"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "interval",
          "candles"
        ],
        "title": "CandlesOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Horizon": {
        "type": "string",
        "enum": [
          "30m",
          "1h",
          "4h",
          "24h"
        ],
        "title": "Horizon",
        "description": "Tahmin ufku. Değerler API ve DB'de olduğu gibi kullanılır."
      },
      "HorizonStateOut": {
        "properties": {
          "horizon": {
            "type": "string",
            "title": "Horizon"
          },
          "label_tr": {
            "type": "string",
            "title": "Label Tr"
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/PredictionOut"
            },
            "type": "array",
            "title": "Predictions"
          }
        },
        "type": "object",
        "required": [
          "horizon",
          "label_tr",
          "predictions"
        ],
        "title": "HorizonStateOut"
      },
...
      "Interval": {
        "type": "string",
        "enum": [
          "1m",
          "5m",
          "15m",
          "1h",
          "4h",
          "1d"
        ],
        "title": "Interval",
        "description": "Mum zaman dilimi. Değerler Binance API'siyle ve DB ile aynı."
      },
...
      "MarketStateOut": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "price": {
            "$ref": "#/components/schemas/PriceOut"
          },
          "coverage": {
            "items": {
              "$ref": "#/components/schemas/CandleCoverageOut"
            },
            "type": "array",
            "title": "Coverage"
          },
          "horizons": {
            "items": {
              "$ref": "#/components/schemas/HorizonStateOut"
            },
            "type": "array",
            "title": "Horizons"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "price",
          "coverage",
          "horizons"
        ],
        "title": "MarketStateOut"
      },
...
      "OutcomeOut": {
        "properties": {
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Resolved At"
          },
          "price_at_target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price At Target"
          },
          "realized_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Return"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "up",
              "down",
              "unresolved"
            ],
            "title": "Outcome"
          },
          "hit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit"
          },
          "brier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brier"
          },
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          }
        },
        "type": "object",
        "required": [
          "resolved_at",
          "price_at_target",
          "realized_return",
          "outcome",
          "hit",
          "brier",
          "resolved_by"
        ],
        "title": "OutcomeOut"
      },
      "PredictionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "horizon": {
            "type": "string",
            "title": "Horizon"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "target_at": {
            "type": "string",
            "format": "date-time",
            "title": "Target At"
          },
          "price_at": {
            "type": "number",
            "title": "Price At"
          },
          "p_up": {
            "type": "number",
            "title": "P Up"
          },
          "expected_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Low"
          },
          "expected_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected High"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "confidence_label": {
            "type": "string",
            "enum": [
              "low",
              "mid",
              "high"
            ],
            "title": "Confidence Label"
          },
          "conflict": {
            "type": "boolean",
            "title": "Conflict"
          },
          "veto_active": {
            "type": "boolean",
            "title": "Veto Active"
          },
          "veto_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veto Reason"
          },
          "source": {
            "type": "string",
            "enum": [
              "live",
              "baseline",
              "backtest"
            ],
            "title": "Source"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "non_overlapping": {
            "type": "boolean",
            "title": "Non Overlapping"
          },
          "report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutcomeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "signals": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SignalOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signals"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "horizon",
          "as_of",
          "target_at",
          "price_at",
          "p_up",
          "expected_low",
          "expected_high",
          "confidence",
          "confidence_label",
          "conflict",
          "veto_active",
          "veto_reason",
          "source",
          "model_version",
          "non_overlapping",
          "report",
          "outcome"
        ],
        "title": "PredictionOut"
      },
      "PredictionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PredictionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "PredictionPage"
      },
      "PriceOut": {
        "properties": {
          "last": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last"
          },
          "change_24h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change 24H"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          }
        },
        "type": "object",
        "required": [
          "last",
          "change_24h",
          "as_of",
          "stale"
        ],
        "title": "PriceOut"
      },
      "SignalOut": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "coverage": {
            "type": "number",
            "title": "Coverage"
          },
          "components": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Components"
          },
          "rationale": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "type": "object",
        "required": [
          "module",
          "score",
          "confidence",
          "coverage",
          "components",
          "rationale"
        ],
        "title": "SignalOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}